bootconf: remove --update and --update-other mode
We haven't been used in the last year if at all. The only known use-case
for them was the custom KDE applet kupdate, which has been decommissioned
in the early beta stage.

In practical sense this hadn't gotten much meaningful testing, so let's
remove it so users don't accidentally brick their systems.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>