bootconf: deprecate and remove update options
Neither of the update options have been used since the early alpha
stages. To prevent accidental misuse and users bricking their systems
let's remove them.

As an added bonus - the "update-disable" option wasn't even documented.

Preserve all the deprecated options in the parser, but keep them as-is
for backwards compatibility.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>