Use conf path, not volume label, to construct a fallback boot label
We used to use the volume label which made sense when the config
was on an image specific partition (efi-a, efi-B, etc) but now
the config resides on the shared ESP partition htis is less useful.

We also stopped labelling the efi-A/B volumes uniquely at some
point so the label just said "efi" which was not helpful.

We now extract the image label form the config file path
which has to be unique as the configs all live in the
same directory.

This is only the fallback boot label - we prefer the explicit
'title' from the boot config if it has been set.