Script to toggle the menu-display option for the STeamOS bootloader
This wraps efibootmgr and identifies the default EFI boot entry,
adding the display-menu argument to it if it is absent, or removing
it if it is present.

It also allows the user to toggle the menu option for a specific
boot entry (but it should do the right thing when run without any
arguments).