Make the canary debug calls part of debug.c
The "canary" debug calls are a pair of calls we make early on
to test the debug infrastructure is working.

One tests the WARN_STATUS pathway, the other tests the ERROR_… macros.
If these two messages are ever missing from the debug log it indicates
that logging isn't working.

These used to be called explicitly from efi_main - they're part of the
debug log initialisation now.