Reimplement strnarrow and strwiden with new UTF-* support
We had a simplistic implementation of strwiden and strnarrow that
assumed only limited character support - this was fine when we were
restricted to the built-in UEFI console font.

Now that we can use glyphs from arbitrary PFF2 fonts loaded at runtime
we should also support the full range of unicode code points.