Use the UTF-16 decoder to when rendering text to graphics mode
We used to used to treat each CHAR16 directly as a UINT32 codepoint
(which is correct for the unicode BMP, which is all we used to
support).

However as our strnarrow/strwiden functions now handle UTF-8 and
UTF-16 in their entirety it is possible that boot labels from
our config files have characters outside the BMP in them. Using the
same decoder as strwiden/strnarrow allows us to cope with these.