Fixed a bug in the font renderer near the right screen edge
The blit buffer width for a font is derived from the maximum width
of any character in that font (this is correct, in that you never have
to reallocate the blit buffer for that font after initialising it).

However when manipulating glyph pixmaps (either when drawing into
the blit buffer or blitting them to the framebuffer) we should use
the glyph's device width, which is usually much smaller (and never
bigger).

The device width of a glyph is what's used to calculate the rendered
dimensions of a string so that's what we should _also_ use when
blitting to the framebuffer.

tl;dr - if we use the font's max width to blit then the blitter
calculates that it will overrun the framebuffer boundary when it
does not need to do so, and refuses to complete the operation.