gfx/menu.c: choose mode picks a UI video mode dynamically
We used to hardwire video mode #3 for boot menu presentation as
this was the only landscape mode on the deck. We can't do this any
more as we're extending support to other devices.

We implement a scoring mechanism that picks the "best" mode,
preferring portrait modes over landscape ones. This should work on
devices as follows:

device                         | outcome
-------------------------------+-------------------------------|----
portrait only                  | bigger, wider modes preferred | OK
landscape only                 | ditto (eg legion mode #0)     | OK
mixed-mode, natively landscape | ditto (eg deck mode #3)       | OK
mixed-mode, natively portrait  | It Depends                    | ???

On a mixed-mode, natively portrait device: If the portrait modes
are of a similar size to the landscape ones we'll probably end up
with a sideways menu.