steamcl-install: handle HD entries as well as /File ones
steamcl-install inspects the UEFI boot entries to figure out if there
is already a suitable UEFI boot variable for the chainloader.

This commit allows it to understand HD boot entries as well as
File style ones (without it you could end up with duplicate boot
entries if the existing entry was an HD one).